+++ modules/main-api/src/main/java/com/whoz_in/main_api/command/member/application/command/LogOutHandler.java
@@ -1,6 +1,6 @@
 package com.whoz_in.main_api.command.member.application.command;
 
-import com.whoz_in.main_api.command.member.application.token.RefreshTokenStore;
+import com.whoz_in.main_api.shared.jwt.tokens.RefreshTokenStore;
 import com.whoz_in.main_api.command.shared.application.CommandHandler;
 import com.whoz_in.main_api.shared.application.Handler;
 import lombok.RequiredArgsConstructor;
