+++ modules/main-api/src/main/java/com/whoz_in/main_api/command/member/application/command/ReissueHandler.java
@@ -2,7 +2,7 @@ package com.whoz_in.main_api.command.member.application.command;
 
 import com.whoz_in.domain.member.model.AccountType;
 import com.whoz_in.domain.member.model.MemberId;
-import com.whoz_in.main_api.command.member.application.token.RefreshTokenStore;
+import com.whoz_in.main_api.shared.jwt.tokens.RefreshTokenStore;
 import com.whoz_in.main_api.command.shared.application.CommandHandler;
 import com.whoz_in.main_api.shared.application.Handler;
 import com.whoz_in.main_api.shared.jwt.tokens.AccessToken;
